refactor: #66 Config 분리
