+++ src/main/java/org/thisway/security/config/AuthenticationConfig.java
@@ -1,0 +1,50 @@
+package org.thisway.security.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.thisway.security.service.CustomUserDetailsService;
+
+@Configuration
+public class AuthenticationConfig {
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    /**
+     * 이 메서드는 현재 어떻게 작동되는지 이해하지 못하고 넣어둠...
+     */
+    @Bean
+    AuthenticationEntryPoint restAuthEntryPoint() {
+        return (request, response, authException) -> {
+            response.sendError(HttpStatus.UNAUTHORIZED.value(), authException.getMessage());
+        };
+    }
+
+    @Bean
+    public DaoAuthenticationProvider daoAuthenticationProvider(
+            CustomUserDetailsService userDetailsService,
+            PasswordEncoder passwordEncoder
+
+    ) {
+        DaoAuthenticationProvider provider = new DaoAuthenticationProvider();
+        provider.setUserDetailsService(userDetailsService);
+        provider.setPasswordEncoder(passwordEncoder);
+        return provider;
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(
+            AuthenticationConfiguration authenticationConfiguration
+
+    ) throws Exception {
+        return authenticationConfiguration.getAuthenticationManager();
+    }
+}
